requirement.txt
asgiref==3.5.2
Django==3.2.13
django-cors-headers==3.12.0
djangorestframework==3.13.1
mysqlclient==2.1.0
passlib==1.7.4
PyMySQL==1.0.2
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.2.0
